Move --transformers option validation to parse-command-line-options.js
All options are now nicely handled in one place.  The rest of the
program can rely on all the options being valid.
